feat: add support for registering and calling type-safe custom MCP methods
